feat(eisland-windows-processes-attacker): expand module functionality with additional source files
- Added multiple source files (`string_utils.c`, `process_ops.c`, `napi_helpers.c`, `napi_binding.c`) to enhance the capabilities of the Windows processes attacker module.
- Updated the module entry point to register new N-API exports, improving the overall functionality and maintainability of the codebase.
